Migrate to using core module
