+++ dependencies.gradle
@@ -4,6 +4,8 @@ repositories {
 		name = "Modmaven"
 		url = "https://modmaven.dev"
 		content {
+			// Tesseract
+			includeGroup "net.swedz"
 			// GrandPower
 			includeGroup "dev.technici4n"
 			// Modern Industrialization
@@ -29,6 +31,13 @@ repositories {
 }
 
 dependencies {
+	jarJar(api("net.swedz:tesseract-api:${tesseract_version}")) {
+		version {
+			strictly tesseract_version_range
+			prefer tesseract_version
+		}
+	}
+
 	if (project.runtime_include_emi.toBoolean()) {
 		localImplementation "dev.emi:emi-neoforge:${emi_version}"
 	} else {
@@ -38,10 +47,8 @@ dependencies {
 	if (project.runtime_include_mi.toBoolean()) {
 		localImplementation "aztech:Modern-Industrialization:${modern_industrialization_version}"
 		localImplementation "dev.technici4n:GrandPower:${grandpower_version}"
-		localImplementation "me.shedaniel.cloth:cloth-config-neoforge:${cloth_config_version}"
 	} else {
 		compileOnly "aztech:Modern-Industrialization:${modern_industrialization_version}"
 		compileOnly "dev.technici4n:GrandPower:${grandpower_version}"
-		compileOnly "me.shedaniel.cloth:cloth-config-neoforge:${cloth_config_version}"
 	}
 }
