+++ gradle.properties
@@ -35,10 +35,10 @@ mod_description=Another library mod for mod developers
 mod_github=https://github.com/Swedz/tesseract-neoforge
 
 # Dependencies
-mixinextras_version=0.3.5
+tesseract_version=1.2.0
+tesseract_version_range=[1.2.0-, 1.3-)
 emi_version=1.1.10+1.21
 modern_industrialization_version=2.4.0
 modern_industrialization_version_range=[2.4.0-, 2.5-)
 grandpower_version=3.0.0
-cloth_config_version=15.0.127
 ohthetreesyoullgrow_version_range=[5.0.7,)
