--- src/main/java/net/swedz/tesseract/neoforge/api/tuple/Pair.java
@@ -1,5 +1,0 @@
-package net.swedz.tesseract.neoforge.api.tuple;
-
-public record Pair<A, B>(A a, B b)
-{
-}
