--- src/main/java/net/swedz/tesseract/neoforge/api/tuple/Triple.java
@@ -1,5 +1,0 @@
-package net.swedz.tesseract.neoforge.api.tuple;
-
-public record Triple<A, B, C>(A a, B b, C c)
-{
-}
