+++ src/main/java/net/swedz/tesseract/neoforge/compat/mi/TesseractMI.java
@@ -2,7 +2,7 @@ package net.swedz.tesseract.neoforge.compat.mi;
 
 import net.minecraft.network.chat.Style;
 import net.swedz.tesseract.neoforge.Tesseract;
-import net.swedz.tesseract.neoforge.api.Assert;
+import net.swedz.tesseract.api.Assert;
 import net.swedz.tesseract.neoforge.compat.mi.component.craft.multiplied.EuCostTransformer;
 import net.swedz.tesseract.neoforge.compat.mi.hook.MIHooks;
 import net.swedz.tesseract.neoforge.compat.mi.hook.context.listener.HatchMIHookContext;
