+++ src/main/java/net/swedz/tesseract/neoforge/compat/mi/machine/builder/MachineBuilder.java
@@ -4,7 +4,7 @@ import aztech.modern_industrialization.machines.MachineBlock;
 import aztech.modern_industrialization.machines.models.MachineCasing;
 import aztech.modern_industrialization.machines.recipe.MachineRecipeType;
 import com.google.common.collect.Lists;
-import net.swedz.tesseract.neoforge.api.Assert;
+import net.swedz.tesseract.api.Assert;
 import net.swedz.tesseract.neoforge.compat.mi.hook.MIHook;
 import net.swedz.tesseract.neoforge.compat.mi.machine.builder.function.MachineBlockEntityWithGuiFactory;
 import net.swedz.tesseract.neoforge.compat.mi.machine.builder.function.MachineBlockFactory;
