+++ src/main/java/net/swedz/tesseract/neoforge/compat/mi/machine/builder/MachineWithGuiBuilder.java
@@ -2,7 +2,7 @@ package net.swedz.tesseract.neoforge.compat.mi.machine.builder;
 
 import aztech.modern_industrialization.compat.rei.machines.SteamMode;
 import aztech.modern_industrialization.machines.recipe.MachineRecipeType;
-import net.swedz.tesseract.neoforge.api.Assert;
+import net.swedz.tesseract.api.Assert;
 import net.swedz.tesseract.neoforge.compat.mi.hook.MIHook;
 import net.swedz.tesseract.neoforge.compat.mi.machine.builder.function.MachineGuiConfigurator;
