--- src/main/java/net/swedz/tesseract/neoforge/config/ConfigManager.java
@@ -1,195 +1,0 @@
-package net.swedz.tesseract.neoforge.config;
-
-import com.google.common.collect.Lists;
-import com.google.common.collect.Sets;
-import net.neoforged.neoforge.common.ModConfigSpec;
-import net.swedz.tesseract.neoforge.api.Assert;
-import net.swedz.tesseract.neoforge.config.annotation.ConfigComment;
-import net.swedz.tesseract.neoforge.config.annotation.ConfigKey;
-import net.swedz.tesseract.neoforge.config.annotation.ConfigOrder;
-import net.swedz.tesseract.neoforge.config.annotation.Range;
-import net.swedz.tesseract.neoforge.config.annotation.SubSection;
-import net.swedz.tesseract.neoforge.config.exception.IllegalConfigMethodException;
-import net.swedz.tesseract.neoforge.helper.NamingConventionHelper;
-import net.swedz.tesseract.neoforge.interfaceproxy.InterfaceProxyManager;
-import net.swedz.tesseract.neoforge.serialization.TomlOps;
-
-import java.lang.reflect.InvocationHandler;
-import java.lang.reflect.Method;
-import java.lang.reflect.Proxy;
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
-
-public final class ConfigManager extends InterfaceProxyManager.WithArgument<ConfigHandler, ConfigManagerArg>
-{
-	private final ConfigCodecMap codecs = new ConfigCodecMap();
-	
-	private boolean includeDefaultValueComments;
-	
-	public ConfigCodecMap codecs()
-	{
-		return codecs;
-	}
-	
-	public ConfigManager includeDefaultValueComments()
-	{
-		includeDefaultValueComments = true;
-		return this;
-	}
-	
-	@Override
-	protected ConfigManagerArg defaultManagerArgument()
-	{
-		return new ConfigManagerArg(Optional.empty(), "");
-	}
-	
-	private <P> ModConfigSpec defaultConfigSpec(Class<P> proxyClass)
-	{
-		try
-		{
-			var builder = new ModConfigSpec.Builder();
-			this.buildSpec(builder, proxyClass);
-			return builder.build();
-		}
-		catch(Throwable ex)
-		{
-			throw new RuntimeException(ex);
-		}
-	}
-	
-	@Override
-	protected <P> ConfigHandler createHandler(Class<P> proxyClass, ConfigManagerArg arg)
-	{
-		return new ConfigHandler(this, arg.spec().orElse(this.defaultConfigSpec(proxyClass)), arg.path());
-	}
-	
-	@Override
-	protected <P> ConfigInstance<P> createInstance(Class<P> proxyClass, P proxy, ConfigHandler handler)
-	{
-		return new ConfigInstance<>(proxyClass, proxy, handler);
-	}
-	
-	@Override
-	public <P> ConfigInstance<P> build(Class<P> proxyClass)
-	{
-		return (ConfigInstance<P>) super.build(proxyClass);
-	}
-	
-	@Override
-	public <P> ConfigInstance<P> build(Class<P> proxyClass, ConfigManagerArg arg)
-	{
-		return (ConfigInstance<P>) super.build(proxyClass, arg);
-	}
-	
-	private <C> void buildSpec(ModConfigSpec.Builder builder, Class<C> configClass) throws Throwable
-	{
-		Assert.noneNull(builder, configClass);
-		
-		var proxy = Proxy.newProxyInstance(configClass.getClassLoader(), new Class[]{configClass}, new DefaultValueConfigHandler());
-		
-		List<Method> methods = Lists.newArrayList(configClass.getMethods());
-		methods.sort((a, b) ->
-		{
-			if(a.isAnnotationPresent(ConfigOrder.class) && b.isAnnotationPresent(ConfigOrder.class))
-			{
-				var orderA = a.getAnnotation(ConfigOrder.class).value();
-				var orderB = b.getAnnotation(ConfigOrder.class).value();
-				return Integer.compare(orderA, orderB);
-			}
-			return 0;
-		});
-		
-		Set<String> keys = Sets.newHashSet();
-		
-		for(var method : methods)
-		{
-			if(method.isAnnotationPresent(ConfigKey.class))
-			{
-				if(method.getReturnType() == void.class)
-				{
-					if(method.getParameterCount() != 1)
-					{
-						throw new IllegalConfigMethodException("Cannot have config editing method with not exactly 1 parameter");
-					}
-					continue;
-				}
-				
-				if(method.getParameterCount() != 0)
-				{
-					throw new IllegalConfigMethodException("Cannot have config method with parameters");
-				}
-				
-				if(method.isAnnotationPresent(ConfigComment.class))
-				{
-					String[] comments = method.getAnnotation(ConfigComment.class).value();
-					builder.comment(comments);
-				}
-				
-				String key = method.getAnnotation(ConfigKey.class).value();
-				if(key.isEmpty())
-				{
-					key = NamingConventionHelper.fromCamelCaseToSnakeCase(method);
-				}
-				Class type = method.getReturnType();
-				
-				if(!keys.add(key))
-				{
-					throw new IllegalConfigMethodException("Duplicate config key: %s".formatted(key));
-				}
-				
-				if(method.isAnnotationPresent(SubSection.class))
-				{
-					builder.push(key);
-					buildSpec(builder, type);
-					builder.pop();
-					continue;
-				}
-				
-				if(method.isDefault())
-				{
-					var defaultValue = InvocationHandler.invokeDefault(proxy, method);
-					this.defineValue(method, builder, key, defaultValue);
-				}
-				else
-				{
-					throw new IllegalConfigMethodException("Cannot retrieve default value from method %s".formatted(method.getName()));
-				}
-			}
-		}
-	}
-	
-	private void defineValue(Method method,
-							 ModConfigSpec.Builder builder, String key,
-							 Object defaultValue) throws Throwable
-	{
-		Class type = method.getReturnType();
-		
-		Object value;
-		if(codecs.has(type))
-		{
-			var codec = codecs.get(type);
-			value = codec.encodeStart(TomlOps.INSTANCE, defaultValue)
-					.getOrThrow((error) -> new IllegalConfigMethodException("Unable to encode default value %s: %s".formatted(defaultValue, error)));
-			builder.define(key, value, (currentValue) -> codec.parse(TomlOps.INSTANCE, currentValue).isSuccess());
-		}
-		else
-		{
-			if(includeDefaultValueComments)
-			{
-				builder.comment("Default: " + defaultValue);
-			}
-			value = defaultValue;
-			if(Range.maybeDefine(builder, key, value, method))
-			{
-				return;
-			}
-			if(value instanceof Enum enumValue)
-			{
-				builder.defineEnum(key, enumValue);
-				return;
-			}
-			builder.define(key, value);
-		}
-	}
-}
