--- src/main/java/net/swedz/tesseract/neoforge/config/ConfigManagerArg.java
@@ -1,12 +1,0 @@
-package net.swedz.tesseract.neoforge.config;
-
-import net.neoforged.neoforge.common.ModConfigSpec;
-
-import java.util.Optional;
-
-public record ConfigManagerArg(
-		Optional<ModConfigSpec> spec,
-		String path
-)
-{
-}
