--- src/main/java/net/swedz/tesseract/neoforge/config/annotation/ConfigComment.java
@@ -1,13 +1,0 @@
-package net.swedz.tesseract.neoforge.config.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface ConfigComment
-{
-	String[] value();
-}
