+++ src/main/java/net/swedz/tesseract/neoforge/helper/datagen/FontDatagenProvider.java
@@ -16,7 +16,7 @@ import net.minecraft.data.PackOutput;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.PackType;
 import net.neoforged.neoforge.common.data.ExistingFileHelper;
-import net.swedz.tesseract.neoforge.api.Assert;
+import net.swedz.tesseract.api.Assert;
 
 import java.nio.file.Path;
 import java.util.Collections;
