+++ src/main/java/net/swedz/tesseract/neoforge/helper/guigraphics/GuiGraphicsBatch.java
@@ -10,7 +10,7 @@ import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.client.renderer.ShaderInstance;
 import net.minecraft.resources.ResourceLocation;
-import net.swedz.tesseract.neoforge.api.Assert;
+import net.swedz.tesseract.api.Assert;
 import org.joml.Matrix4f;
 
 import java.util.function.Consumer;
