+++ src/main/java/net/swedz/tesseract/neoforge/helper/guigraphics/TesseractGuiGraphics.java
@@ -13,7 +13,7 @@ import net.minecraft.client.renderer.texture.MissingTextureAtlasSprite;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.FastColor;
 import net.minecraft.util.FormattedCharSequence;
-import net.swedz.tesseract.neoforge.api.Assert;
+import net.swedz.tesseract.api.Assert;
 import org.jetbrains.annotations.ApiStatus;
 import org.joml.Matrix4f;
