--- src/main/java/net/swedz/tesseract/neoforge/interfaceproxy/InterfaceProxyEntry.java
@@ -1,6 +1,0 @@
-package net.swedz.tesseract.neoforge.interfaceproxy;
-
-public interface InterfaceProxyEntry<R>
-{
-	R resolve(Object[] args);
-}
