+++ src/main/java/net/swedz/tesseract/neoforge/lang/LangManager.java
@@ -16,9 +16,9 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.material.Fluid;
-import net.swedz.tesseract.neoforge.api.Assert;
+import net.swedz.tesseract.api.Assert;
 import net.swedz.tesseract.neoforge.api.WorldPos;
-import net.swedz.tesseract.neoforge.interfaceproxy.InterfaceProxyManager;
+import net.swedz.tesseract.interfaceproxy.InterfaceProxyManager;
 import net.swedz.tesseract.neoforge.lang.annotation.PlaceholderKey;
 import net.swedz.tesseract.neoforge.lang.exception.UndefinedParserException;
 import net.swedz.tesseract.neoforge.lang.exception.UndefinedStyleException;
